Consolidate all constants into src/constants.rs
- Move all constants from src/protocol/constants.rs to src/constants.rs under new stateless_proxy module
- Update all imports to use crate::constants::stateless_proxy::*
- Remove src/protocol/constants.rs and clean up src/protocol/mod.rs
- Fix STREAMING_CHUNK_SIZE magic number (65536) to use constant
- All 148 tests pass
