+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "nntp-proxy"
-version = "0.1.0"
+version = "0.2.0"
 edition = "2024"
 description = "High-performance NNTP proxy server with connection pooling and authentication"
 license = "MIT"
