+++ src/proxy.rs
@@ -13,7 +13,7 @@ use tracing::{debug, error, info, warn};
 use crate::config::{Config, ServerConfig};
 use crate::network::SocketOptimizer;
 use crate::pool::{BufferPool, ConnectionProvider, DeadpoolConnectionProvider, PoolPrewarmer};
-use crate::protocol::*;
+use crate::constants::stateless_proxy::*;
 use crate::router;
 use crate::session::ClientSession;
 use crate::types;
